EzGrind/Backend/requirements.txt
flask
flask-cors
mysql-connector-python
werkzeug
